+++ test/PublishOption.Tests.ps1
@@ -23,7 +23,7 @@ InModuleScope azure.datafactory.tools {
             { Get-Command -Name New-AdfPublishOption -ErrorAction Stop } | Should -Not -Throw
         }
 
-        Context 'When called' {
+        Context 'When called without args' {
             It 'Should return object of AdfPublishOption type' {
                 $script:result = New-AdfPublishOption
                 $script:result.GetType() | Should -Be 'AdfPublishOption'
@@ -40,8 +40,35 @@ InModuleScope azure.datafactory.tools {
                 $script:result.DeleteNotInSource | Should -Be $false
                 $script:result.StopStartTriggers | Should -Be $true
             }
+        }
+        
+        Context 'When called with wrong FilterFilePath' {
+            It 'Should throw exception' {
+                {
+                    New-AdfPublishOption -FilterFilePath "this-file-does-not-exist.fji3ugf4.txt"
+                } | Should -Throw 
+            }
+        }
+
+        Context 'When called with correct FilterFilePath' {
+            It 'Should not throw exception' {
+                {
+                    $script:opt = New-AdfPublishOption -FilterFilePath ".\test\BigFactorySample2\deployment\filter.option1.txt"
+                } | Should -Not -Throw 
+            }
+            It 'Should contains 2 rules added to Includes with appropriate values' {
+                $script:opt.Includes.Count | Should -Be 2
+                $script:opt.Includes.ContainsKey('pipeline.*') | Should -Be $true
+                $script:opt.Includes.ContainsKey('trigger.*') | Should -Be $true
+            }
+            It 'Should contains 2 rules added to Excludes with appropriate values' {
+                $script:opt.Excludes.Count | Should -Be 2
+                $script:opt.Excludes.ContainsKey('*.SharedIR*') | Should -Be $true
+                $script:opt.Excludes.ContainsKey('*.LS_SqlServer_DEV19_AW2017') | Should -Be $true
+            }
 
         }
+
         
     } 
 }